requirements.txt
async-generator==1.10
attrs==22.1.0
backcall==0.2.0
beautifulsoup4==4.11.1
behave==1.2.6
cachetools==5.2.0
certifi==2022.6.15
cffi==1.15.1
chardet==5.0.0
charset-normalizer==2.1.1
colorama==0.4.5
commonmark==0.9.1
cryptography==37.0.4
cssselect==1.1.0
decorator==5.1.1
execnet==1.9.0
fasteners==0.17.3
filelock==3.8.0
google-api-core==2.8.2
google-api-python-client==2.58.0
google-auth==2.11.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.56.4
h11==0.13.0
httplib2==0.20.4
idna==3.3
iniconfig==1.1.1
ipdb==0.13.9
ipython==7.34.0
jedi==0.18.1
matplotlib-inline==0.1.6
more-itertools==8.14.0
nose==1.3.7
oauth2client==4.1.3
outcome==1.2.0
packaging==21.3
parameterized==0.8.1
parse==1.19.0
parse-type==0.6.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==2.5.2
pluggy==1.0.0
prompt-toolkit==3.0.30
protobuf==4.21.5
ptyprocess==0.7.0
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyDrive==1.3.1
Pygments==2.13.0
PyMySQL==1.0.2
pyOpenSSL==22.0.0
pyotp==2.6.0
pyparsing==3.0.9
PySocks==1.7.1
pytest==7.1.2
pytest-forked==1.4.0
pytest-html==2.0.1
pytest-metadata==2.0.2
pytest-ordering==0.6
pytest-rerunfailures==10.2
pytest-xdist==2.5.0
python-dotenv==0.20.0
PyYAML==6.0
qbittorrent==0.1.6
requests==2.28.1
rich==12.5.1
rsa==4.9
sbvirtualdisplay==1.1.0
selenium==4.4.3
seleniumbase==4.2.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
toml==0.10.2
tomli==2.0.1
traitlets==5.3.0
trio==0.21.0
trio-websocket==0.9.2
uritemplate==4.1.1
urllib3==1.26.12
wcwidth==0.2.5
wsproto==1.2.0
